requirements.txt
click
pydantic
fastapi
uvicorn
pytest
python-jose[cryptography]
